v1/microservice-3/historical_transaction_mq/requirements.txt
pika==1.2.0
requests
python-dotenv